Deactivate Pixmap exchange
Sporadicly the content of panel popups is not displayed when doing Pixmap
exchanges. Deactivate it therefore for now as a workaround.

This means that at worst we always blit the picture. But the performance
impact seems to be negligible with Xrender compositing. With GL ES we can
still do buffer flips, what is faster anyways.